Add colors and icons
For handling the colors of DWM I've chosen to add a folder for themes.
To create a new theme simply make a file in the 'themes' folder, e.g.
'colortheme.h' and source it in config.h under // themes. I've also
added icons for tags cause it's like super super cool 8).